Move non-testable failing action tests to StepHandler test
